Add Travis CI integration
Performs the usual tests, and also performs auto-generation of docs
on CI success.

Docs are located at: 'https://tiehuis.github.io/tetrs/tetrs/'.

Note: This commit is a squashed commit and force overwrites over remote.
We consider this ok since the repo is just created (and no-one is likely
using it).

The number of commits is due to testing live against Travis.