Abstract rotation data behind a trait
This allows us to implement a number of seperate rotational system
more easily and better compartmentalizes the internal data.

This is closes #1 for the most part. We could make some small additions
to this by taking the static `Block` method instances and tying those
to a rotation system instead. This makes more sense from a user
perspective, and these arguments require a rotation system as an
argument anyway.