Allow `Block` to take general RotationSystem
In order to do this in a manner which provides a clean API,
`RotationSystem` implementors return a static reference to an instance.

This is fine at the moment, since `RotationSystem`'s have no members
anyway. This idea can likely be transferred to the `Wallkick` trait
implementation as well.

This leaves only `Randomizer`'s to determine their implementation in the
`Engine`. This likely will still use trait objects, but may be
complicated due to the using of non-static instances.