Fix block/field logic errors
This fixes some incorrect calculations in the functions to clear lines
and check piece collisions.

The field internal data was changed from `(x, y)` ordering to `(y, x)`.
This makes working with the individual rows far easier.

A schema test is temporarily omitted due to these changes. This is
likely a problem in how the schema works, and not the field/block.