Add frontend sub-crate for testing purposes
This provides a minimal implementation over the engine provided by
tetrs. This will likely continue to grow and have more features, purely
for the sake of engine integration testing.

This frontend will not be a complete function game, and an external
package which utilizes this library would instead be better suited.