Add options struct which stores game engine options
These could be placed directly in a game field, but it really pollutes
the entries and makes it much less clear what is important and what
isn't.

Options should hold only primitive types, mostly flags, and only things
which matter internally, i.e. no ghost enabled flag or preview count
flags, which are dealt with at the application level for the most part.