Add lag detection and add initial logging support
This adds some lag detection in the update phase which is useful when
building a new interface. There are a number of reasons an update could
be lagging, such as incorrect time calculations or a too intensive draw
phase. This helps debug this far easier.

The Cargo.lock file was removed as well. It is considered god practice
to have it for binaries, but since this is more a library with the
binary only as a testing portion, it is considered unnecessary.