Update internal types and add 'safe' conversion macros
The internal type casting can get a little messy, as mentioned in #2.
It is difficult to not use `usize` directly for some things, due to the
nature they are used (accessing arrays a lot).

For now, we only make minor modifications and replace all casts with
macros which can be adjusted in the future to provide safer casting
(i.e. no negative i32 casts to usize, assert!(x > 0)).