Replace Builder initialization with Default-based
Instead of using method syntax to modify an object, we now rely on
specific 'Options' structs which can be used to constructed certain
structs.

An example:

  let field = Field::with_options(FieldOptions {
                    width: 15, height: 20,
                    ..Default::default()
              })

This can be slightly more cumbersome in some cases, but otherwise
reduces writing special method cases for new parameters, and removes the
need to modify an existing object.

Also added is the ability to load configuration from a file. This only
applies to the engine, but any 'Options' struct is simple enough that it
supports serialization (serde is used with all currently).

Another change is made in adding factory functions to wallkick and
rotation_system modules. This makes initialization from a file much
easier.

Finally, a new module named `import` has been added which reexports
modules in a standard way. This simplifies importing of all the
submodules a lot.