Rewrite engine.rs to have better internal management
The main engine is probably the most difficult part of this to write. It
involves a lot of stateful changes which make it very difficult to keep
track of.

This version is a fairly simplified version, however it provides a lot
of customisable support already and lays out a good structure with which
new status modes and the like can be bolted on. Many ideas were derived
from the internal loop of NullpoMino, however it is cleaned up and
simplified drastically in the current case (extra features
withstanding).

There are still a number of bugs to fix within this, and things such as
History and Statistics are not utilized, however the base is here and
works moderately well.